requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
opencv-python>=4.5.5
albumentations>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
pillow>=9.0.0
matplotlib>=3.5.0
tensorboard>=2.9.0